@@ -17,7 +17,7 @@ class _ChartApp extends StatelessWidget {
 
 class _MyHomePage extends StatefulWidget {
   // ignore: prefer_const_constructors_in_immutables
-  _MyHomePage({Key key}) : super(key: key);
+  _MyHomePage({Key? key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -124,9 +124,9 @@ class _MyHomePageState extends State<_MyHomePage> {
 class ChartSampleData {
   ChartSampleData({this.x, this.open, this.high, this.low, this.close});
 
-  final DateTime x;
-  final double open;
-  final double high;
-  final double low;
-  final double close;
+  final DateTime? x;
+  final double? open;
+  final double? high;
+  final double? low;
+  final double? close;
 }
